refactor(format): extract redaction extension policy
